% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makePlotURI}
\alias{makePlotURI}
\title{Helper function to make the ggplot output from LIME suitable for leaflet.}
\usage{
makePlotURI(expr, width, height, ...)
}
\description{
Helper function to make the ggplot output from LIME suitable for leaflet.
}
