% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explainable_sdm.R
\name{plot_explainable_sdm}
\alias{plot_explainable_sdm}
\title{Plot interactive explainable SDM map}
\usage{
plot_explainable_sdm(input_data, processed_plots,
  use_selected_feature = FALSE)
}
\arguments{
\item{input_data}{A DataFrame containing the lime explanations and data point coordinates.}

\item{processed_plots}{A list of plots processed for leaflet map.}

\item{use_selected_feature}{A logical indicating if a feature is extracted for plotting.}
}
\value{
Leaflet map
}
\description{
Create an interactive map, showing the feature explanations for every observation.
}
\examples{
\dontrun{
# build and train a machine learning model
task <- makeClassifTask(id = "model", data = data.train, target = "label")
lrn <- makeLearner("classif.lda", predict.type = "prob")
mod <- train(lrn, task)

# prepare explainable data
explainable_data <- prepare_explainable_data(data.test, mod, coordinates.df)

# process lime plots to make them suitable for leaflet
processed_plots <- process_lime_plots(explainable_data$explanation)

# plot the interactive leaflet map
plot_explainable_sdm(explainable_data$processed_data, processed_plots)
}
}
