% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_explainable_data.R
\name{prepare_explainable_data}
\alias{prepare_explainable_data}
\title{Explain SDM models}
\usage{
prepare_explainable_data(test_dataset, mlr_model, coordinates_df,
  selected_feature = "none", randomize = FALSE,
  randomize_proportion = 0.1)
}
\arguments{
\item{test_dataset}{A DataFrame containing the test data.}

\item{mlr_model}{A trained MLR model}

\item{coordinates_df}{A Dataframe containing the coordinates.}

\item{selected_feature}{A character indicating if a feature should be extracted for plotting.}

\item{randomize}{Boolean deciding if a sample of the dataset should be taken}

\item{randomize_proportion}{Float deciding what proportion of the data should be sampled}
}
\value{
A list containing data necessary for the interactive map
}
\description{
Prepare data for plotting interactive explainable SDM.
}
\examples{
\dontrun{
# build and train a machine learning model
task <- makeClassifTask(id = "model", data = data.train, target = "label")
lrn <- makeLearner("classif.lda", predict.type = "prob")
mod <- train(lrn, task)

# prepare explainable data
explainable_data <- prepare_explainable_data(data.test, mod, coordinates.df)
}
}
