% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_lime_plots.R
\name{process_lime_plots}
\alias{process_lime_plots}
\title{Process LIME plots}
\usage{
process_lime_plots(explanation)
}
\arguments{
\item{explanation}{DataFrame containing LIME explained data}
}
\value{
list containing processed plots
}
\description{
Process LIME plots to make them usable in an interactive map
}
\examples{
\dontrun{
# build and train a machine learning model
task <- makeClassifTask(id = "model", data = data.train, target = "label")
lrn <- makeLearner("classif.lda", predict.type = "prob")
mod <- train(lrn, task)

# prepare explainable data
explainable_data <- prepare_explainable_data(data.test, mod, coordinates.df)

# process lime plots to make them suitable for leaflet
processed_plots <- process_lime_plots(explainable_data$explanation)
}
}
