% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmexplain-package.r
\docType{package}
\name{sdmexplain}
\alias{sdmexplain}
\alias{sdmexplain-package}
\title{sdmexplain: Making Species Distribution Models more explainable}
\description{
Species Distribution Models, like other machine learning technique tend to operate as black boxes. Despite our best efforts to measure how accurate they work, often we (and domain experts even more so) lack trust in the prediction. Recent research has provided us with the tools necessary to improve this. `sdmexplain` helps by integrating those tools.
}
