% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explainable_sdm.R
\name{plot_explainable_sdm}
\alias{plot_explainable_sdm}
\title{Plot interactive explainable SDM map}
\usage{
plot_explainable_sdm(explanation_coordinates, map_df, processed_plots)
}
\arguments{
\item{explanation_coordinates}{DataFrame containing the coordinates of the data points.}

\item{map_df}{DataFrame containing lime explanations data.}

\item{processed_plots}{list List of plots processed for leaflet map}
}
\value{
Leaflet map
}
\description{
Create an interactive map, showing the feature explanations for every observation.
}
\examples{
\dontrun{
# build and train a machine learning model
task <- makeClassifTask(id = "model", data = data.train, target = "label")
lrn <- makeLearner("classif.lda", predict.type = "prob")
mod <- train(lrn, task)

# prepare explainable data
explainable_data <- prepare_explainable_data(data.test, mod, coordinates.df)

# process lime plots to make them suitable for leaflet
processed_plots <- process_lime_plots(explainable_data$explanation)

# plot the interactive leaflet map
plot_explainable_sdm(explainable_data$explanation_coordinates,
explainable_data$map_df,
processed_plots)
}
}
