% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_explainable_data.R
\name{prepare_explainable_data}
\alias{prepare_explainable_data}
\title{Explain SDM models}
\usage{
prepare_explainable_data(test_dataset, mlr_model, coordinates_df)
}
\arguments{
\item{test_dataset}{DataFrame containing the test data}

\item{mlr_model}{Trained MLR model}
}
\value{
list containing data necessary for the interactive map
}
\description{
Prepare data for plotting interactive explainable SDM.
}
\examples{
\dontrun{
# build and train a machine learning model
task <- makeClassifTask(id = "model", data = data.train, target = "label")
lrn <- makeLearner("classif.lda", predict.type = "prob")
mod <- train(lrn, task)

# prepare explainable data
explainable_data <- prepare_explainable_data(data.test, mod, coordinates.df)
}
}
